pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "packmin"
version = "0.1.0"
description = "AI-powered minimalist packing list generator using capsule wardrobe principles"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Min" }
]
keywords = ["travel", "packing", "ai", "minimalist", "capsule-wardrobe"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "requests>=2.28",
    "openai>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
]

[project.scripts]
packmin = "packmin.cli:main"

[project.urls]
Homepage = "https://github.com/minhsao/packmin"
Repository = "https://github.com/minhsao/packmin"

[tool.hatch.build.targets.wheel]
packages = ["packmin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
